Autenticação usando Laravel Sanctum - 3º Commit

+++ api/app/Http/Controllers/PostController.php
@@ -5,10 +5,18 @@ namespace App\Http\Controllers;
 use App\Models\Post;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 
-
-class PostController extends Controller
+class PostController extends Controller implements HasMiddleware
 {
+    public static function middleware()
+    {
+        return[
+            new Middleware('auth:sanctum', except:['index','show'])
+        ];
+    }
     /**
      * Display a listing of the resource.
      */
@@ -27,7 +35,7 @@ class PostController extends Controller
             'body' => 'required'
         ]);
 
-        $post = Post::create($fields);
+        $post = $request->user()->posts()->create($fields);
 
         return $post;
     }
@@ -46,23 +54,27 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        Gate::authorize('modify', $post);
+        
         $fields = $request->validate([
             'title' => 'required|max:225',
             'body' => 'required'
         ]);
-
+        
         $post->update($fields);
-
+        
         return $post;
     }
-
+    
     /**
      * Remove the specified resource from storage.
-     */
+    */
     public function destroy(Post $post)
     {
+        Gate::authorize('modify', $post);
+         
         $post->delete();
-
+        
         return [ 'massage' => 'The post was deleted'];
     }
 }
